refactor: [SRTP-302] move SSL related class into their own package
