+++ src/main/java/it/gov/pagopa/rtp/activator/configuration/ssl/DefaultSslContextFactory.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.rtp.activator.configuration;
+package it.gov.pagopa.rtp.activator.configuration.ssl;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
