+++ src/main/java/it/gov/pagopa/rtp/activator/configuration/ssl/EnvVarsSslContextPropsProvider.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.rtp.activator.configuration;
+package it.gov.pagopa.rtp.activator.configuration.ssl;
 
 import java.util.Objects;
 import org.springframework.lang.NonNull;
