+++ src/main/java/it/gov/pagopa/rtp/activator/configuration/ssl/SslContextPropsProvider.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.rtp.activator.configuration;
+package it.gov.pagopa.rtp.activator.configuration.ssl;
 
 /**
  * Defines a provider for SSL context properties.
