+++ src/test/java/it/gov/pagopa/rtp/activator/configuration/ssl/DefaultSslContextFactoryTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.rtp.activator.configuration;
+package it.gov.pagopa.rtp.activator.configuration.ssl;
 
 import static org.junit.jupiter.api.Assertions.*;
